+++ CHANGELOG.md
@@ -1,2 +1,4 @@
 # Changelog
+* (April 18th 2024). Tested with WordPress 6.5.2.
+* (April 18th 2024). Added a dashboard page.
 * (October 14th 2020). Initial Release.